debugs

+++ nw.h
@@ -1,1 +1,1 @@
-int nw(char *s1, char *s2, int match_score, int mismatch_score, int gap_score, int print_alignments);
+double nw(char *s1, char *s2, int match_score, int mismatch_score, int gap_score, int print_alignments);
